Implement QAT for APoT
[ghstack-poisoned]
