+++ test/quantization/core/experimental/quantization_util.py
@@ -1,10 +1,10 @@
 import torch
-import torch.nn as nn
 import torchvision
 import torchvision.transforms.transforms as transforms
 import os
 import torch.quantization
 from torchvision.models.quantization.resnet import resnet18
+from torch.autograd import Variable
 
 # Setup warnings
 import warnings
@@ -19,11 +19,14 @@ warnings.filterwarnings(
 )
 
 """
-Define helper functions
+Define helper functions for APoT PTQ and QAT
 """
 
 # Specify random seed for repeatable results
 _ = torch.manual_seed(191009)
+
+train_batch_size = 30
+eval_batch_size = 50
 
 class AverageMeter(object):
     """Computes and stores the average and current value"""
@@ -34,7 +37,7 @@ class AverageMeter(object):
 
     def reset(self):
         self.val = 0
-        self.avg = 0
+        self.avg = 0.0
         self.sum = 0
         self.count = 0
 
@@ -128,127 +131,25 @@ def prepare_data_loaders(data_path):
 
     return data_loader, data_loader_test
 
-data_path = '~/my_imagenet/'
-
-train_batch_size = 30
-eval_batch_size = 50
-
-data_loader, data_loader_test = prepare_data_loaders(data_path)
-criterion = nn.CrossEntropyLoss()
-float_model = resnet18(pretrained=True)
-float_model.eval()
-
-# deepcopy the model since we need to keep the original model around
-import copy
-model_to_quantize = copy.deepcopy(float_model)
-
-model_to_quantize.eval()
-
-"""
-Prepare models
-"""
-
-# Note that this is temporary, we'll expose these functions to torch.quantization after official releasee
-from torch.quantization.quantize_fx import prepare_fx, convert_fx
-
-def calibrate(model, data_loader):
-    model.eval()
-    with torch.no_grad():
-        for image, target in data_loader:
-            model(image)
-
-from torch.ao.quantization.experimental.qconfig import (
-    uniform_qconfig_8bit,
-    apot_weights_qconfig_8bit,
-    apot_qconfig_8bit,
-    uniform_qconfig_4bit,
-    apot_weights_qconfig_4bit,
-    apot_qconfig_4bit
-)
-
-"""
-Prepare full precision model
-"""
-full_precision_model = float_model
-
-top1, top5 = evaluate(full_precision_model, criterion, data_loader_test)
-print("Model #0 Evaluation accuracy on test dataset: %2.2f, %2.2f" % (top1.avg, top5.avg))
-
-"""
-Prepare model PTQ for specified qconfig for torch.nn.Linear
-"""
-def prepare_ptq_linear(qconfig):
-    qconfig_dict = {"object_type": [(torch.nn.Linear, qconfig)]}
-    prepared_model = prepare_fx(copy.deepcopy(float_model), qconfig_dict)  # fuse modules and insert observers
-    calibrate(prepared_model, data_loader_test)  # run calibration on sample data
-    return prepared_model
-
-"""
-Prepare model with uniform activation, uniform weight
-b=8, k=2
-"""
-
-prepared_model = prepare_ptq_linear(uniform_qconfig_8bit)
-quantized_model = convert_fx(prepared_model)  # convert the calibrated model to a quantized model
-
-top1, top5 = evaluate(quantized_model, criterion, data_loader_test)
-print("Model #1 Evaluation accuracy on test dataset (b=8, k=2): %2.2f, %2.2f" % (top1.avg, top5.avg))
-
-"""
-Prepare model with uniform activation, uniform weight
-b=4, k=2
-"""
-
-prepared_model = prepare_ptq_linear(uniform_qconfig_4bit)
-quantized_model = convert_fx(prepared_model)  # convert the calibrated model to a quantized model
-
-top1, top5 = evaluate(quantized_model1, criterion, data_loader_test)
-print("Model #1 Evaluation accuracy on test dataset (b=4, k=2): %2.2f, %2.2f" % (top1.avg, top5.avg))
-
-"""
-Prepare model with uniform activation, APoT weight
-(b=8, k=2)
-"""
-
-prepared_model = prepare_ptq_linear(apot_weights_qconfig_8bit)
-
-top1, top5 = evaluate(prepared_model, criterion, data_loader_test)
-print("Model #2 Evaluation accuracy on test dataset (b=8, k=2): %2.2f, %2.2f" % (top1.avg, top5.avg))
-
-"""
-Prepare model with uniform activation, APoT weight
-(b=4, k=2)
-"""
-
-prepared_model = prepare_ptq_linear(apot_weights_qconfig_4bit)
-
-top1, top5 = evaluate(prepared_model, criterion, data_loader_test)
-print("Model #2 Evaluation accuracy on test dataset (b=4, k=2): %2.2f, %2.2f" % (top1.avg, top5.avg))
-
-
-"""
-Prepare model with APoT activation and weight
-(b=8, k=2)
-"""
-
-prepared_model = prepare_ptq_linear(apot_qconfig_8bit)
-
-top1, top5 = evaluate(prepared_model, criterion, data_loader_test)
-print("Model #3 Evaluation accuracy on test dataset (b=8, k=2): %2.2f, %2.2f" % (top1.avg, top5.avg))
-
-"""
-Prepare model with APoT activation and weight
-(b=4, k=2)
-"""
-
-prepared_model = prepare_ptq_linear(apot_qconfig_4bit)
-
-top1, top5 = evaluate(prepared_model, criterion, data_loader_test)
-print("Model #3 Evaluation accuracy on test dataset (b=4, k=2): %2.2f, %2.2f" % (top1.avg, top5.avg))
-
-"""
-Prepare eager mode quantized model
-"""
-eager_quantized_model = resnet18(pretrained=True, quantize=True).eval()
-top1, top5 = evaluate(eager_quantized_model, criterion, data_loader_test)
-print("Eager mode quantized model evaluation accuracy on test dataset: %2.2f, %2.2f" % (top1.avg, top5.avg))
+def training_loop(model, criterion, data_loader):
+    optimizer = torch.optim.Adam(model.parameters(), lr=0.001)
+    train_loss, correct, total = 0, 0, 0
+    count = 0
+    model.train()
+    for i in range(10):
+        for data, target in data_loader:
+            # if count == 1:
+            #     break
+            print(count)
+            count += 1
+            optimizer.zero_grad()
+            output = model(data)
+            loss = criterion(output, target)
+            loss = Variable(loss, requires_grad=True)
+            loss.backward()
+            optimizer.step()
+            train_loss += loss.item()
+            _, predicted = torch.max(output, 1)
+            total += target.size(0)
+            correct += (predicted == target).sum().item()
+    return train_loss, correct, total
